Imlementing game play with database

+++ Your.Melody.Library/Helpers/IPointsCounter.cs
@@ -4,6 +4,6 @@ namespace Your.Melody.Library.Helpers
 {
     public interface IPointsCounter
     {
-        Task<double> CountingPointsAsync(Song song, string titleByUser, string artistByUser, double secWhenUserResponce);
+        Task<float> CountingPointsAsync(Song song, string titleByUser, string artistByUser, float secWhenUserResponce);
     }
 }
